ternyata gini punya app docker

+++ server/internal/app/products/handler.go
@@ -35,17 +35,10 @@ func (handler *ProductHandlerImpl) FindAll(ctx *fiber.Ctx) error {
 	products, err := handler.Service.FindAll(customContext)
 
 	if err != nil {
-		return ctx.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
-			"status":  "failed to get products",
-			"message": err.Error(),
-		})
+		exception.ResponseError(ctx, fiber.StatusNotFound, fiber.ErrNotFound.Message, err)
 	}
 
-	// return resultnya
-	return ctx.Status(fiber.StatusOK).JSON(fiber.Map{
-		"status":  "success",
-		"message": products,
-	})
+	return helper.Response(ctx, "successfully findall", products)
 
 }
 
@@ -56,15 +49,15 @@ func (handler *ProductHandlerImpl) FindByID(ctx *fiber.Ctx) error {
 	productId, err := uuid.Parse(ctx.Params("productId"))
 
 	if err != nil {
-		exception.ResponseError(ctx, fiber.StatusNotFound, "Not Found Product", err)
+		return exception.ResponseError(ctx, fiber.StatusNotFound, "Not Found Product", err)
 	}
 
 	product, err := handler.Service.FindByID(customContext, productId)
 	if err != nil {
-		exception.ResponseError(ctx, fiber.StatusInternalServerError, "Failed to find Product", err)
+		return exception.ResponseError(ctx, fiber.StatusInternalServerError, "Failed to find Product", err)
 	}
 
-	return helper.Response(ctx, product)
+	return helper.Response(ctx, "successfully find id", product)
 }
 
 func (handler *ProductHandlerImpl) Create(ctx *fiber.Ctx) error {
@@ -73,25 +66,17 @@ func (handler *ProductHandlerImpl) Create(ctx *fiber.Ctx) error {
 	defer cancel()
 
 	if err := ctx.BodyParser(&product); err != nil {
-		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
-			"status":  "bad request",
-			"message": err.Error(),
-		})
+		return exception.ResponseError(ctx, fiber.StatusBadRequest, fiber.ErrBadRequest.Message, err)
 	}
 
 	// Create one product
 	product, err := handler.Service.Save(customContext, product)
 	if err != nil {
-		return ctx.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
-			"status":  "failed to create product",
-			"message": err.Error(),
-		})
+		return exception.ResponseError(ctx, fiber.StatusInternalServerError, fiber.ErrInternalServerError.Message, err)
+
 	}
 
-	return ctx.Status(fiber.StatusCreated).JSON(fiber.Map{
-		"status":  "successfully created product",
-		"message": product,
-	})
+	return helper.Response(ctx, "Successfully created product", product)
 }
 
 func (handler *ProductHandlerImpl) Edit(ctx *fiber.Ctx) error {
@@ -101,57 +86,39 @@ func (handler *ProductHandlerImpl) Edit(ctx *fiber.Ctx) error {
 	defer cancel()
 
 	if err := ctx.BodyParser(&product); err != nil {
-		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
-			"status":  "bad request",
-			"message": err.Error(),
-		})
+		return exception.ResponseError(ctx, fiber.StatusBadRequest, fiber.ErrBadRequest.Message, err)
 	}
 
 	productId, err := uuid.Parse(ctx.Params("productId"))
 
 	if err != nil {
-		return ctx.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
-			"status":  "Internal server error",
-			"message": err.Error(),
-		})
+		return exception.ResponseError(ctx, fiber.StatusInternalServerError, fiber.ErrInternalServerError.Message, err)
 	}
 
 	result, err := handler.Service.Update(customContext, productId, product)
 
 	if err != nil {
-		return ctx.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
-			"status":  "failed to update product",
-			"message": err.Error(),
-		})
+		return exception.ResponseError(ctx, fiber.StatusInternalServerError, fiber.ErrInternalServerError.Message, err)
 	}
 
-	return ctx.Status(fiber.StatusOK).JSON(fiber.Map{
-		"status":  "success",
-		"message": result,
-	})
+	return helper.Response(ctx, "Successfully created product", result)
 
 }
 
 func (handler *ProductHandlerImpl) Destroy(ctx *fiber.Ctx) error {
 	customContext, cancel := context.WithCancel(context.Background())
-
 	defer cancel()
 
+	// Parsing UUID dari parameter URL
 	productId, err := uuid.Parse(ctx.Params("productId"))
-	if err != nil {
-		return ctx.Status(fiber.StatusNotFound).JSON(fiber.Map{
-			"status":  "ID Not Found",
-			"message": err.Error(),
-		})
+	if err != nil || productId == uuid.Nil {
+		return exception.ResponseError(ctx, fiber.StatusNotFound, fiber.ErrNotFound.Message, err)
 	}
 
+	// Menghapus produk
 	err = handler.Service.Delete(customContext, productId)
-
 	if err != nil {
-		return ctx.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
-			"status":  "failed to delete product",
-			"message": err.Error(),
-		})
+		return exception.ResponseError(ctx, fiber.StatusInternalServerError, fiber.ErrInternalServerError.Message, err)
 	}
 
 	return ctx.Status(fiber.StatusOK).JSON(fiber.Map{
